+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## 2020-01-07
+
+### Added
+
+- Allow for filtering by source tag on dataset admin listing pages
+
+
 ## 2019-12-27
 
 ### Added